No longer build dependencies by default, add --include-dependencies / -d
Most of the time when I use e.g. `cheribuild.py cheribsd` I don't want
it to also build LLVM, so add a short flag to achieve that and deprecate
-t flag as this behaviour is now the default.

Pseudo-targets such as "all" will still build all their dependencies by
default. The implementation for this is quite hacky but something better