No longer require projects to be added to targets.py manually
After reading some stuff on python metaclasses I realized that they can
be used to automatically create a Target instance whenever a new Project
subclass is defined. Not sure this makes the code any more understandable
but at least now adding support for building new projects is easier and
the dependencies for a project are specified in the same file as the
steps needed to build the project.