Properly handle stashing changes if repository has git submodules
This problem happens with QEMU now that we checkout submodules as well.
`git diff` outputs something like
```
diff --git a/dtc b/dtc
--- a/dtc
+++ b/dtc
@@ -1 +1 @@
-Subproject commit 65cc4d2748a2c2e6f27f1cf39e07a5dbabd80ebf
+Subproject commit 65cc4d2748a2c2e6f27f1cf39e07a5dbabd80ebf-dirty
```

But if we run git stash We get `No local changes to save` and the following
`git stash pop` will fail. We need to check for this message and skip the
pop step if that happens.