Fix running symlinks to cheribuild.py with python 3.6
I have a symlink to cheribuild.py in ~/bin and for some reason python
only looks for module imports in the directory where the symlink is
and not the resolved file. This seems to be a change in python 3.6
as https://hg.python.org/cpython/rev/47c31e7d3779 suggests this should
work as expected.