Allow running cheribuild inside a docker container
There is a flag --docker to enable this behaviour or alternatively you can
run the provided `docker-cheribuild.py` symlink. When running the symlink
cheribuild will read the configuration from
`~/.config/docker-cheribuild.json` and determine which directories should
 be used for the sources/build/output. Currently all other config options
are ignore in the container since the config file does not exist there.
To work around this I could probably also map the cheribuild config file
into the container but that may cause other issues.

The source/build/output and cheribuild directories from the host are mapped
into the docker container and then cheribuild.py is invoked inside the
container.

Also added a Dockerfile that can be used for running this. In the future
we probably want to use a container that includes CHERI clang/llvm instead
so that we don't have to build those again