Build gdb-native as part of the sdk
This required a bit of refactoring since gdb depended on sdk (even though
this is only the case for the MIPS version not the host one). To debug
dependency loops like this I added the new --print-targets-only flag
which prints which targets will be executed but unlike --pretend doesn't
print the commands that will be executed.