Silence some messages that should only be printed when building that target
Now that I end up calling __init__ even when not building the targets some
of them may print messages when they shouldn't. Even when building the
target this is confusing since __init__ is called for all classes before
the first process() so it is not clear which target this message belongs to.