Write mtree files directly when building disk-image and stop using makefs -F
This means we no longer need to copy the extra files for the rootfs into
the rootfs install dir but can instead use the mtree contents= key

This requires parsing the METALOG and adding our own entries instead. It
also has the advantage that we no longer need to use the -F flag of makefs
which is probably wrong but happens to work in most cases.

Makefs without the -F flag is a bit more selective on the mtree keys that
it accepts so for now we need to strip the tags= and time= key from all
entries.