Avoid duplicate mtree directory definitions (causes makefs failures
The makefs for cheribsd was failing (if you included the CheriBSD tests)
because the mtree contained the following lines:
`./usr/lib/debug//usr/tests` and then later on `./usr/lib/debug/usr/tests`.
This caused makefs to complain that there were duplicate definitions of
`tests` (but it only reports the duplicate and not where it was defined
before).