Make test_metalog work with python 3.4 to keep Jenkins happy
Not sure it makes sense to keep python 3.4 support around much longer. It
was originally needed because that was the version installed on the FreeBSD
build servers but all of our systems should have at least python 3.5 now.

I will probably drop support for it the next time I need some newer API.