Add a cheribsd-purecap target that builds world as purecap
cheribsd still defaults to building cheribsd as a hybrid world (this is
the cheribsd-cheri target). I might change this in the future and instead
add a cheribsd-hybrid target and keep make cheribsd an alias for
cheribsd-purecap.

cheribsd-purecap will build a shared world by default which can be changed
to static by speicifying --cheribsd-purecap/build-static