Make `cheribuild.py libcxx --test` work for CHERI too
Previously it would only work for native and sometimes for CHERI (if QEMU
was running on the right port). Now I spawn a minimal cheribsd instance
just to run the tests and override the libcxx executor.

This will also allow spawning e.g. 16 QEMU instances and distributing the
tests across those instances.