Add "bsdtar" target to install bsdtar on machines without root access
It is needed in order to build the sysroot archive. GNU tar is not compatible
with the command that is currently used. I'm not sure it's the best way of
building the sysroot archive, but that's the way it was done since
build_sdk.sh and changing it might add/remove some files from the archive.