Make failed commands fail the tests
Before the tests only failed if we got a CHERI trap or a known error
message. Now append
`;if test $? -eq 0; then echo '__COMMAND' 'SUCCESSFUL__'; else echo '__COMMAND' 'FAILED__'; fi`
to check the exit code of the running command and throw an exception if
we don't get either of the two values before timing out.