gdb: also set all configure env vars during make step
It appears that the gdb build runs ./configure scripts lazily (i.e. during
the make stage rather than running them all from the top-level ./configure)
In this case PKG_CONFIG_SYSROOT_DIR is not set and therefore the
configure scripts looks for libsource-highligh on the host. This causes
the build to fail if source-highlight is installed.