Improve adding runtime linker(s) to minimal disk images
The lack of ld-elf.so was caused by an incorrect change to CheriBSD.
We need it to run any binaries. However, instead of hardcoding the list
of runtime linkers in a .files list, we check for the presence of all
ld-elf.so and all possible ld-*.so compat ABI runtime linkers.
This ensures that ld-elf64.so1. will be added for systems where CheriABI
is the default ABI instead of checking for ld-cheri-elf.so.1

This mosftly reverts commit dde5d21ae6241cae74f59149a7a5be0cce853a96 and
d8c7a146f6db62c81d51dbbe3a70fc4eba4e2708