Fix handling of full_rebuild_if_older_than
We only need to check the date of the last clean operation, not when the
last build completed. Before this change, a failed build was not writing
the timestamp file so any following builds would perform a clean build
again. This is not required: all we need to ensure is that the build
directory has been deleted, not that a build completed after this date.