Default to building tests for CheriBSD
If you want to reduce build times and don't care about the tests it's still
possible to set `{ "cheribsd": { "build-tests": false }}` in
~/.config/cheribuild.json or override it on the command line.

Now that FreeBSD installs kyua as part of the base install there is no need
to default tests to off since they can be run easily.
This default made more sense when we didn't have kyua since it was really
complicated to install (and took about 30 minutes).

Change suggested by @trasz