qemu-mount-rootfs.sh: Also add symlinks for the other rootfs dirs
On a hybrid or purecap CheriBSD instance, qemu-mount-rootfs.sh will now
also mount the non-cheri rootfs under /nfsroot-nocheri, the hybrid one at
/nfsroot-hybrid and the purecap on at /nfsroot-purecap.

This works as long  as all these directories are a subdir of the output
root (i.e. the default), but won't work if the user explicitly specifies
a different dir for the rootfs outside the output root. In that case the
script will simply create an invalid symlink.