Add support for building+running CheriBSD for AArch64
The tricky thing here is creating the disk image: some versions of makefs
are broken and can't create the FAT filesystem correctly. I upstreamed
fixes for that but if you are stuck on an older version you can change the
`use_makefs = True` line to fall back to using https://github.com/vapier/mtools.git
instead.