Remove -stdlib=libc++ from the essential compiler flags
This used to be required because upstream Clang defaulted to using
libstdc++ unless you used a freebsd triple with at least version 10. This
meant that a plain mips64-unknown-freebsd tried to build against libstdc++
which did not exist. CHERI Clang has defaulted to using libc++ for FreeBSD
purecap and FreeBSD without a version triple for quite a while so this
workaround should no longer be needed.

This should fix an error reported by @brettferdosi:
Adding -stdlib=libc++ unconditionally can also break some autotools checks
that run with -Werror since it will trigger a -Wunused-command-line-argument
warning that is promoted to an error.