filesystemutils.py: Add rewrite_file/add_unique_line_to_file helpers
These will be used to add bash to /etc/shells but could be useful more
generally for arbitrary in-place file modifications. These hide and
slightly munge line endings (always put one at the end of the file even
if not originally present) since this gives a nicer interface, but does
mean there may be cases where they're not appropriate.