disk_image.py: Use master.passwd in rootfs
It's always been installed to etc/ in the rootfs, but where it lived in
the source tree moved. There's no good reason not to use the installed
copy, only downsides in case other targets want to modify master.passwd
(or the upstream build system wants to dynamically configure it, or even
just move it again).