bash.py: Add option to set root's shell to bash
This is generally frowned upon in FreeBSD, since bash isn't part of the
base system. But we're not using ports and upgrading in the same way, so
that issue is gone. Normally you'd also log in as a non-root user that
has bash as the login shell, but we (ab)use root as it's just an
ephemeral VM. But, as always, caveat emptor.