Add --sysroot-install-dir configuration option
cheribuild currently installs the sysroots under the sdk dir, which is derived
from the tools_root, which, in turn, defaults to the output_dir.  When
attempting to share tools directories between long-lived branches of cheribsd
(for example), this derivation chain implies that the sysroots end up in the
nominally shared tools directory, which is not desirable.  This change allows
cheribuild to still use a shared compiler and qemu (and so on) and yet use
different sysroots.