Use argparse.parse_intermixed_args()
This results in better error messages than the current approach of using
parse_known_args() and appending all unknown arguments to the targets.
Now argparse will print the --help message for all arguments that start
with a dash rather than adding them to the list of targets.

The reason I used the parse_known_args() hack is to allow command lines
such as `cheribuild.py target1 --arg1 target2` but it turns out
parse_intermixed_args() exists exactly for that purpose.