Stop defaulting to hybrid/purecap MIPS for targets such as run/cheribsd/etc.
We support many different target architectures and while it made sense a
few years ago to default to hybrid/purecap MIPS for all of these targets,
we should be explicit now.

This patch changes (almost) all targets that can be built for more than
one architecture to require an explicit suffix.
The only targets that retain a default mapping are llvm -> llvm-native
(since users probably want to build a native CHERI llvm and not an
experimental version build for MIPS/RISC-V) and the FETT targets which
default to purecap CHERI-RISC-V.

Note to self: the code to select targets and the metaclass that adds the
per-architecture classes dynamically is an awful accident caused by small
incremental changes and should really be rewritten.