Default to building FreeBSD with the system LLVM installation
This means we no longer have to build a version of upstream LLVM by
default and more importantly means we no longer attempt to build with a
snapshot of the LLVM master branch that may have new warnings that break
the -Werror build. This only really makes sense if you intend to fix those
issues or want to see if FreeBSD builds with the latest LLVM snapshot.

The old behaviour can still be enabled by setting --freebsd/toolchain to
"default-external"/"upstream-llvm" (or "cheri-llvm" if you want to use
an existing CHERI LLVM installation instead of upstream LLVM).