Remove QEMU from disk-image dependencies
It's only required if you want to build QCOW2 images instead of the
default raw image. The other command that's used (qemu-img) is only needed
to show some debug information about the image. Therefore it's safe to
remove this dependency. I noticed this since we shouldn't need QEMU to
build a Morello image.