CheriBSD tests: stop when we enter an inifite panic loop
This is currently happening when running the full test suite on RISC-V
(it appears to be triggered by on of the sys/netinet tests). However, we
weren't detecting "panic: Fatal page fault at 0x" as one of the panic()
messages so the tests just kept running until the entire job timed out.