Fix Morello FVP version detection
The latest FVP is version 0.11 rather than 0.10 and the rev number is now
lower than what we previously checked for. This will cause us to not
enable virtio and try to add the morello plugin (which no longer exists
since it is embedded in the FVP now).
Fix this by parsing the --version output instead of reading