Add gdb-*-purecap targets that build hybrid gdb for a purecap rootfs
This avoids having cherbsd-*-hybrid as a dependency of run-*-purecap just
to get a sysroot for building GDB. While the target name is a bit
confusing, using -purecap to build a hybrid gdb simplifies downstream
consumers. This also includes some refactoring in target_info to allow
building a against a purecap rootfs but using the compiler+compile flags
for a hybrid (or non-CHERI) build.