pexpect test scripts: add a --disable-coredumps option
When set, coredumps are disabled on startup. This is useful to speed up
the full testsuite job in jenkins since there are quite a few tests that
crash and writing the coredumps takes a long time. The default behaviour
is to keep coredumps enabled.

Also change the default coredump directory to /tmp if there is no usable
smbfs mount to avoid filling up the disk (minimal images only a have a few
megabytes of free space).