Drop failok from non-minimal fstabs for /
If this fails something has gone horribly wrong. The presence of the
option causes issues when booting from loader(8) since the kernel does
not understand the option, causing it to fail and retry without options,
presumably with the root device taken from the kernel config, which
works but is noisy.