u-boot: Install as QEMU firmware and provide method to get the path
Whilst we don't get the same benefit of -bios where the directory is
automatically searched, this is logically where the file belongs and
ensures it's not forgotten when creating archives. If we really wanted
to in future we could build a bbl with u-boot embedded in it and use
that as -bios to be closer to how real EFI firmware works, but that
seems unnecessary for now. Besides, if we were doing this properly we'd
use EDK2 with its embedded OpenSBI as the -bios.