Include the OS name in target names when not CheriBSD or native
We default to building for CheriBSD, so building for CheriBSD does not add
the os (e.g. sqlite-riscv64-purecap not sqlite-cheribsd-riscv64-purecap).
However, for all other operating systems we do want to include the suffix.

This change should be NFC but will allow building e.g. gdb-freebsd-amd64
in the future.