Handle being executed from a non-interactive shell script
When run from an interactive shell, a new process group is created for
us (so job control works) and we are made the terminal leader (since
that is what foreground processes expect). However, if invoked from a
non-interactive shell (or something just forks and execs us), we go and
create a process group for ourselves but don't change the leader of the
terminal (although we do that for executing some children like QEMU,
which probably would result in us becoming the leader after they exit),
so ^C does not work, and we seem to also end up unable to use input().