u-boot-fett: Add a new variant for using on F1
This is a stripped-down version of the QEMU platform for use with the
Connectal-based F1 setup. AWSteria should be able to use the normal QEMU
one, as the only important difference is the location of DRAM (which is
unfortunately rather hard-coded in various places, including in string
constants for embedded scripts so awkward to parameterise).