Switch to use loader.efi on the UEFI ESP
FreeBSD is moving away from boot1.efi. Follow that by switching to
use one of the loader.efi files when building a UEFI image.

To allow for a minimal image without bootloader configuration files
in the root filesystem use loader_simp.efi there and the default
loader_lua.efi on the full image.