Work around Meson generated CMake toolchain file requiring a string
For some reason CMAKE_C_FLAGS_INIT can't be a CMake list, it needs to be
a space-separated string instead. Since there is no way to convert from
a Meson list to a space-separated string in the toolchain file, we work
around this by adding new @TOOLCHAIN_FOO_STR@ replacements.
With this workaround finding libraries via CMake appears to work.

I've filed a feature request for array->str conversion as
https://github.com/mesonbuild/meson/issues/8534