Fix forcing static linkage with CMake
I believe this used to work in the past, but didn't when I tried building
libxml2 static (it found libz.so instead of libz.a).
The reason appears to be that project() overwrites the value of
CMAKE_FIND_LIBRARY_SUFFIXES that I set in the toolchain file.
As suggested in https://gitlab.kitware.com/cmake/cmake/-/issues/21942#note_921012,
we can use CMAKE_PROJECT_INCLUDE to work around this.