run_fvp: Clean up interactive/x11/headless logic
The names were confusing and code was duplicated. We also needlessly forwarded
X11 on macOS for x11=True interactive=False, not that that combination is used
anywhere. We also still had hard-coded port forwarding for exposing telnet
ports; this should have used find_free_port, but is unnecessary now we have
headless supported automatically, as you will either get xterms started
automatically or we will wrap the FVP and run telnet for you.