Make README more generation even more automated
We have generate-readme-help.sh, but that only generates the output to
put in the code block, it doesn't actually put it there, requiring
something like a visual select and %!./generate-readme-help.sh in vim.
Instead, use some magic comments to mark the region and turn it into a
script that automatically replaces everything between them.