[UPSTREAM?][pexpect] Add option to chunk to chunk sent data
When interacting with FPGAs and/or QEMU, it can happen that sent bytes
are lost if we send long command lines. This commit adds an option to
add a short delay after every N characters written. Setting the chunk
size to 100 has so far allowed me to avoid missing characters when running
QtBase tests on CheriBSD, whereas before we would get some missing chars
in most runs.