Meson: add /usr/local/<arch>/lib to RPATH for CheriBSD
We install most projects to a location that is not part of the default
library search path. CMake and autotools projects automatically set
RPATH to include those directories but meson only does it if the pkgconfig
files explicitly add -Wl,-rpath. Work aruond this by adding to LDFLAGS.

With this change libfontconfig.so has the expected RPATH and Okular can
now be used to view PDF files without setting LD_LIBRARY_PATH.