Test script: mount the build dir under the host path
This is needed for at least the libjpeg-turbo tests. We previously mounted
the build dir under /build and added a symlink but that breaks tests
that try to get at the source directory using a relative path
(../../my-srcdir ends up resolving to /my-srcdir. I believe this is because
/build is used as the CWD).

We could probably also use mount_nullfs to work around this, but mounting
the smbfs file system in the correct location is simpler.