boot_cheribsd: Use stricter, less fragile expects
Currently we reuse the same boot_messages list multiple times, which is
silly, because we should never see a loader(8) prompt after the first
match, and should never see the mounting rootfs prompt after that's been
seen. Instead, split it up into three lists, each a superset of the last
(in case the previous stage isn't present, or is present but isn't
printed, such as bootverbose messages).