boot_cheribsd: Support loader_lua autoboot
This prints a fancier prompt that doesn't match the current string. It
also wants a carriage return as the enter key (i.e. what you get with a
proper serial console, as well as QEMU normally given it puts the tty in
raw mode), not a line feed which will just stop the countdown. Finally,
printing the message that matched is both silly and not very useful,
instead print a nicer message like the other ones here.