Reduce pre-push time by 1/3 by optimizing set_env()
Profiling `cheribuild.py -p __run_everything__ --build-and-test --benchmark -q`
showed that calling set_env was a large proportion of the total runtime.
The problem is that we create a complete copy of the environment variables
on ever one of the 18316 calls to set_env. Optimize this by remembering
the changed variables and using os.putenv()/os.unsetenv() instead.
This reduces the number of putenv calls from 556580 to 24694.
Before: 11.21s (39s total)
After:   0.23s (26s total).