Exit if <foo>-cheribuild.json is missing when using <foo>-cheribuild.py
A prefixed cheribuild.py to get the default config file name would
previously continue with the default values. However, I believe this is
not a good idea as it could silently do the wrong thing. If you really
want to use a prefixed command but no config file, --config-file=/dev/null
can be used as a workaround.