Add the ability to filter the built dependencies with regexes (#221)
When building e.g. KDE programs, I want to build the KDE framework
dependencies but skip targets that take a long time to build such as QtBase
and all the underlying dependencies. This patch lets me do the following:

cheribuild.py <target> -d --skip-sdk "--skip-dependency-filter=qt.*" "--skip-dependency-filter=libx.*"

This avoids lots of build time (especially when combined with --clean or
--reconfigure since we also need to build a native qtbase+qtdeclarative
for the bootstrap tools in KConfig and other frameworks.