vcu118-run: Add --benchmark-config to disable malloc debugging
We track -CURRENT so MALLOC_PRODUCTION is off by default (unless you are
building for MIPS and pass CHERI=128, as a local diff). This option will
create a malloc.conf that disables abort-on-warning (opt.abort) and
junking (opt.junk) in order to not pick up that debugging overhead and
skew benchmark results.