Add new current-directory and current-directory-<type> targets
These are special targets that perform a build where the source
directory is the current working directory and the build directory's
basename is the name of the current directory.

The current-directory-<type> targets are for a specific build system.
The current-directory target is an even more special target that looks
for various build system files and tries to infer what to use; note that
any options must be set on the real tool target still, as it's
implemented by having a dynamically-calculated dependency.