docker-cheribuild: Drop default source/output/build-root directories check
This check is insufficient to stop you shooting your foot off for
various reasons; for example, changing just one of them is enough but
could leave others the same, and this is also checking against the
default value, not what a non-docker cheribuild is currently using, so
if you configure both to the same custom location (such as outside an
NFS-mounted home directory) then it doesn't detect that.

For the upcoming release we want to have docker-cheribuild look as close
to cheribuild as we can. Whilst users should unpack the release
somewhere other than ~/cheri (to avoid conflicts with a from-source
cheribuild, and to ensure that come next release they don't try and
unpack a newer release over an older one and cause chaos), it should be
supported, and this gets in the way of that.