gdb: Pass --disable-sim to configure
Until recently sim would only build a simulator for the target triple,
if supported, but now it respects --enable-targets=all so we end up
building a huge number of simulators. This is slow (and a waste of time
as we don't use them) but also not all simulators build properly for
FreeBSD (ppc has a header include order issue in one file if configure
decides to use gnulib's stdlib.h, as happens for FreeBSD), nor do they
build with our default warning flags (iq2000 has a whole load of missing
prototype warnings that we make errors).

This is not yet needed for CHERI-GDB, but will be needed for future
versions, as well as for the upcoming upstream target.