Add a MercurialRepository implementation
The main mising feature is automatically changing URLs; we can
programmatically detect when it's outdated but hg provides no
programmatic way to change it, and I do not care enough to manually
mangle the hgrc file. Mercurial repos should be rare and changing URLs
should be even rarer so it's just not worth doing.