Rework for-purecap-rootfs target support
Rather than hacking it up in gdb and cmake, push the idea of a separate
rootfs up into CrossCompileTarget. This is needed to allow purecap and
hybrid-for-purecap-rootfs versions of the same target to exist, rather
than only allowing one or the other; GDB now has a dependency on libgmp
and, whilst we don't currently need the purecap version, we shouldn't
forbid it from existing.

Note that previously for-purecap-rootfs targets would install (unless
overridden, like GDB) to <base>/<cpu>-purecap within the purecap rootfs,
which was bogus. They now install to the directory within the rootfs
corresponding to the architecture for which they are built (e.g.
<cpu>-hybrid for hybrid-for-purecap-rootfs targets). This also only
affects CMake, since the only other for-purecap-rootfs target is GDB.