llvm: Forcefully enable zlib compressed debug support
Due to CMake's broken caching of the macOS sysroot directory (broken as
it doesn't detect that it needs to refresh it after an OS update) my
LLVM builds didn't have zlib support enabled and so I've been running
with the -gz and -Wl,--compress-debug-sections=zlib default compiler
options commented out for a while. Thus, force it to be enabled so
errors are printed during the LLVM build in that case, and add it as a
required pkg-config.

Whilst I could add a needs_configure override (and initially did), when
your cache is as messed up as mine was it needs manual intervention
anyway, and the failure mode for having an LLVM built without zlib
support is pretty obvious.