Add a use-x11 option to Qt
This also relaxes the assertion in ConfigOption.__get__ to allow
loading config values without an object instance since it works just fine
as long as we don't have a default value that depends on the object
instance. Since we don't know that about the default when calling __get__,
we can reject the `callable(self.default)` and still allow access for
most config options.