cheribsd: Default to building alternate kernel ABIs
We want it to be easy to get started with purecap kernels, so make sure
people get them by default.

As part of this, reformat the various kernel ABI tests to be exceedingly
uniform and easy to scan through as the changes in line wrapping make
them hard to trawl through and update.