Add hidden command line option to allow running as root
This can be useful when testing behaviour inside docker, but should not
be exposed to users by default to avoid unexpected errors.
I use this to test cheribuild with older python versions using Pycharm's
docker interpreter feature. Unfortunately it's not possible to override
the UID for the docker interpreters in Pycharm, so explicitly adding this
flag to the run configuration allows me to use the GUI debugger when
tracking down Jenkins failures, etc.