libunwind: Always build and install static library
Currently we end up not building and installing (but possibly building
and not installing) it, only the shared one. This looks like it might be
a CMakeLists.txt mismerge downstream, but we should be explicit here
anyway in case defaults ever change, since libcxxrt requires both in
order to be built.