Introduce --run/use-qemu option for greater flexibility
This allows choosing between default (normally CHERI or system, also
Morello special-cased currently), CHERI, Morello, upstream, system or a
custom QEMU when running the given OS, with some sanity checking for
things like trying to run CHERI on an upstream QEMU. This is intended to
mirror --freebsd/toolchain for picking LLVMs.

As part of this, rework the test framework slightly so each test can be
run standalone and the cached QEMU choice can be forgotten each time.