cheri-syzkaller: Adapt to be a hybrid-for-purecap-rootfs target
This is needed to ensure it's still available once hybrid targets are
disabled by default, as well as it being more useful to have in a normal
purecap rootfs.

Also drop the redundant cheribsd dependency, since this actually causes
issues with X-for-Y-rootfs targets; the target manager ends up trying to
find a cheribsd-mips64-hybrid-for-purecap-rootfs target, which makes no
sense. In theory that should be fixed to special-case cheribsd and remap
the target architecture, but in practice there's no point as it's always
an implied dependency.