Disable and hide hybrid targets by default
Config options will still be recognised, as will the target names
themselves, but any attempt to actually build the target will give an
error. For Jenkins the targets are always enabled so we can test them
and avoid flag days in Jenkinsfile.

Note also that hybrid-for-purecap-rootfs targets don't count, since
those all exist to work around targets that can't yet be built as
purecap but are wanted for purecap images, such as GDB.

Closes: #224