Migrate flake8 config to .flake8 config file
This avoids duplication and allows a bare flake8 run by a developer to
automatically use the right config, rather than having to remember to
set the maximum line length or enable more helpful output.

Note that this drops count; it's rather obsolete thanks to statistics
and is annoying in printing a 0 rather than nothing when everything is
conforming.