cheribsd-release: New target for building release media
This requires changes not yet present in dev or main in order to support
NO_ROOT building, building on non-FreeBSD and CheriBSD-specific tweaks
for lib64 and pure-capability kernel dists, and disabling ZFS on CHERI.

Currently only supports morello-purecap as that is the only target it's
worth building installer images for, and riscv64-purecap would require
us to be using EFI and GENERIC kernels rather than direct booting and
hard-coding rootfs devices in the kernel configs like the QEMU one. It
should support aarch64, and possibly morello-hybrid, but neither seem
worth adding.