Revert "qtdeclarative: Make sure we use our own Qt for build tools qmlcachegen"
The QtDeclarative build also builds a native version of qtdeclarative
and installs it to the QT_HOST_BINS directory, so we should not need
a native one. I believe the real problem is that some downstream
projects use the wrong qmlcachegen. Qt modules such as QtQuickControls
use the version in QT_HOST_BINS correctly, but it seems like the KDE
Plasma .qml files are processed using qmlcachegen from $PATH instead of
using QT_HOST_BINS.

This reverts commit 1a81f4559547e6387b64af677be5ac00f82ab3c0.