Revert "Fix building native Spec2006"
The install dir does not match the SDK dir in Jenkins for an unpacked
SDK. This commit may have fixed non-Jenkins native builds, but totally
breaks any Jenkins build. Native builds should probably instead be fixed
by just not overriding the binaries, or perhaps something else, but this
is definitely not the right way.

This reverts commit 966a930c5de04725ebfb34562f34a70eb62b01d8.