Convert the syzkaller targets to use RISC-V
This is completely untested, I just changed all mips64 occurrences to use
the CPUArchitecture value instead. Also converted both targets to use
supported_architectures instead of pretending it's a native build.
We can no longer build CheriBSD for MIPS and I am planning to remove the
CompilationTargets.CHERIBSD_MIPS_* constants so this needs to be updated.