disk-image-minimal: don't include /boot in MFS images
The MFS-root disk-image classes are children of the disk-image-minimal class and
knew to set self.include_boot = False, but then, on x86, aarch64, or Morello,
the generic add_unlisted_files_to_metalog method goes and adds a significant
fraction of /boot anyway, amounting to some 80 MB of material at present.  Gate
this second inclusion of /boot paths on a new class variable, include_boot_more,
and rename the earlier one to include_boot_kernel in preparation for the next
commit.