disk-image: add /include-kernel CLI flag for MFS
The MFS images default to excluding `/boot`, but perhaps you want it after all
(for its symbol or CTF tables, perhaps).  This new flag just includes the kernel
(and anything else mentioned in `files/minimal-image/boot.files`, should that
ever happen) and does not trigger the inclusion of loader and its files.