project: Make add_config_option's shortname actually useful
This option is currently unused. At the moment, it doesn't prefix the
option with the target, so it occupies the global namespace, which
doesn't make sense for project-specific options. Turn it into an altname
that abuses shortname like we already do for some config options; this
will be used for per-project reconfigure, which will mirror config in
having a per-project force-configure alias that abuses shortname.