cheribsd-release/freebsd-release: Fix weirdness with two release directories
By using the bmake-provided Makefile fragments, or the system-provided
Makefile fragments without the in-tree share/mk in the path, we don't
get our src.sys.obj.mk and things get inconsistent (see the comment for
details). This ensures we only have the one object directory and that
it's where it should be in the hierarchy, i.e. looks like a normal
subdirectory; previously it was hacked around by forcing OBJTOP but this
doesn't work quite right in that the top-level make still uses the
OBJROOT/release path without TARGET.TARGET.ARCH.