qemu: Always enable morello-softmmu
Only ancient versions don't have support. Importantly, the current
implementation is buggy, as setup is run before process, so the local
checkout will not be present (or up to date) the first time round and
thus we won't enable morello-softmmu until a later reconfigure. It's
also misleading as the --help output will omit it.