Add a default_source_dir_in_subdir("subdir") helper
This is needed to make ReuseOtherProjectRepository work for e.g. the KDE
frameworks projects. I added a new target there that made use of that
functionality but noticed that it was not actually reusing the source
directory because the override was unconditionally returning
`config.source_root / "kde-frameworks" / project.default_directory_basename`.
To fix this add a new shared helper and an assertion in the project
constructor to avoid problems like this in the future.

This code is extremely fragile and rather ugly, so should probably be
refactored but for now this is the minimal fix.